% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lottery_function.R
\name{giveMeLotteryNumbers}
\alias{giveMeLotteryNumbers}
\title{(please) give me lottery numbers}
\usage{
giveMeLotteryNumbers(start = 1, end = 45, numbers = 6, games = 1)
}
\arguments{
\item{start}{What number is the minimum number?}

\item{end}{What number is the maximum number?}

\item{numbers}{How many numbers do you need?}

\item{games}{How many games do you want to participate in?}
}
\description{
This program makes lottery numbers for you.
Default setting is for Korean Lotto.
}
\examples{
giveMeLotteryNumbers()
giveMeLotteryNumbers(games = 10)
giveMeLotteryNumbers(end=60, numbers=8, games=1)
}
